TST/CLN: organize SparseArray tests
