+++ pandas/tests/arrays/sparse/test_arithmetics.py
@@ -27,10 +27,6 @@ def mix(request):
 
 
 class TestSparseArrayArithmetics:
-
-    _base = np.array
-    _klass = SparseArray
-
     def _assert(self, a, b):
         # We have to use tm.assert_sp_array_equal. See GH #45126
         tm.assert_numpy_array_equal(a, b)
@@ -54,7 +50,7 @@ class TestSparseArrayArithmetics:
         self._assert(result, expected)
 
     def _check_bool_result(self, res):
-        assert isinstance(res, self._klass)
+        assert isinstance(res, SparseArray)
         assert isinstance(res.dtype, SparseDtype)
         assert res.dtype.subtype == np.bool_
         assert isinstance(res.fill_value, bool)
@@ -133,25 +129,25 @@ class TestSparseArrayArithmetics:
                     mark = pytest.mark.xfail(raises=AssertionError, reason="GH#38172")
                     request.node.add_marker(mark)
 
-        values = self._base([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
+        values = np.array([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
 
-        a = self._klass(values, kind=kind, fill_value=fill_value)
+        a = SparseArray(values, kind=kind, fill_value=fill_value)
         self._check_numeric_ops(a, scalar, values, scalar, mix, op)
 
     def test_float_scalar_comparison(self, kind):
-        values = self._base([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
+        values = np.array([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
 
-        a = self._klass(values, kind=kind)
+        a = SparseArray(values, kind=kind)
         self._check_comparison_ops(a, 1, values, 1)
         self._check_comparison_ops(a, 0, values, 0)
         self._check_comparison_ops(a, 3, values, 3)
 
-        a = self._klass(values, kind=kind, fill_value=0)
+        a = SparseArray(values, kind=kind, fill_value=0)
         self._check_comparison_ops(a, 1, values, 1)
         self._check_comparison_ops(a, 0, values, 0)
         self._check_comparison_ops(a, 3, values, 3)
 
-        a = self._klass(values, kind=kind, fill_value=2)
+        a = SparseArray(values, kind=kind, fill_value=2)
         self._check_comparison_ops(a, 1, values, 1)
         self._check_comparison_ops(a, 0, values, 0)
         self._check_comparison_ops(a, 3, values, 3)
@@ -160,11 +156,11 @@ class TestSparseArrayArithmetics:
         # when sp_index are the same
         op = all_arithmetic_functions
 
-        values = self._base([0.0, 1.0, 2.0, 6.0, 0.0, 0.0, 1.0, 2.0, 1.0, 0.0])
-        rvalues = self._base([0.0, 2.0, 3.0, 4.0, 0.0, 0.0, 1.0, 3.0, 2.0, 0.0])
+        values = np.array([0.0, 1.0, 2.0, 6.0, 0.0, 0.0, 1.0, 2.0, 1.0, 0.0])
+        rvalues = np.array([0.0, 2.0, 3.0, 4.0, 0.0, 0.0, 1.0, 3.0, 2.0, 0.0])
 
-        a = self._klass(values, kind=kind, fill_value=0)
-        b = self._klass(rvalues, kind=kind, fill_value=0)
+        a = SparseArray(values, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, kind=kind, fill_value=0)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
     def test_float_same_index_with_nans(
@@ -180,94 +176,94 @@ class TestSparseArrayArithmetics:
         ):
             mark = pytest.mark.xfail(raises=AssertionError, reason="GH#38172")
             request.node.add_marker(mark)
-        values = self._base([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
-        rvalues = self._base([np.nan, 2, 3, 4, np.nan, 0, 1, 3, 2, np.nan])
+        values = np.array([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
+        rvalues = np.array([np.nan, 2, 3, 4, np.nan, 0, 1, 3, 2, np.nan])
 
-        a = self._klass(values, kind=kind)
-        b = self._klass(rvalues, kind=kind)
+        a = SparseArray(values, kind=kind)
+        b = SparseArray(rvalues, kind=kind)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
     def test_float_same_index_comparison(self, kind):
         # when sp_index are the same
-        values = self._base([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
-        rvalues = self._base([np.nan, 2, 3, 4, np.nan, 0, 1, 3, 2, np.nan])
+        values = np.array([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
+        rvalues = np.array([np.nan, 2, 3, 4, np.nan, 0, 1, 3, 2, np.nan])
 
-        a = self._klass(values, kind=kind)
-        b = self._klass(rvalues, kind=kind)
+        a = SparseArray(values, kind=kind)
+        b = SparseArray(rvalues, kind=kind)
         self._check_comparison_ops(a, b, values, rvalues)
 
-        values = self._base([0.0, 1.0, 2.0, 6.0, 0.0, 0.0, 1.0, 2.0, 1.0, 0.0])
-        rvalues = self._base([0.0, 2.0, 3.0, 4.0, 0.0, 0.0, 1.0, 3.0, 2.0, 0.0])
+        values = np.array([0.0, 1.0, 2.0, 6.0, 0.0, 0.0, 1.0, 2.0, 1.0, 0.0])
+        rvalues = np.array([0.0, 2.0, 3.0, 4.0, 0.0, 0.0, 1.0, 3.0, 2.0, 0.0])
 
-        a = self._klass(values, kind=kind, fill_value=0)
-        b = self._klass(rvalues, kind=kind, fill_value=0)
+        a = SparseArray(values, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, kind=kind, fill_value=0)
         self._check_comparison_ops(a, b, values, rvalues)
 
     def test_float_array(self, kind, mix, all_arithmetic_functions):
         op = all_arithmetic_functions
 
-        values = self._base([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
-        rvalues = self._base([2, np.nan, 2, 3, np.nan, 0, 1, 5, 2, np.nan])
+        values = np.array([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
+        rvalues = np.array([2, np.nan, 2, 3, np.nan, 0, 1, 5, 2, np.nan])
 
-        a = self._klass(values, kind=kind)
-        b = self._klass(rvalues, kind=kind)
+        a = SparseArray(values, kind=kind)
+        b = SparseArray(rvalues, kind=kind)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
         self._check_numeric_ops(a, b * 0, values, rvalues * 0, mix, op)
 
-        a = self._klass(values, kind=kind, fill_value=0)
-        b = self._klass(rvalues, kind=kind)
+        a = SparseArray(values, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, kind=kind)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
-        a = self._klass(values, kind=kind, fill_value=0)
-        b = self._klass(rvalues, kind=kind, fill_value=0)
+        a = SparseArray(values, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, kind=kind, fill_value=0)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
-        a = self._klass(values, kind=kind, fill_value=1)
-        b = self._klass(rvalues, kind=kind, fill_value=2)
+        a = SparseArray(values, kind=kind, fill_value=1)
+        b = SparseArray(rvalues, kind=kind, fill_value=2)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
     def test_float_array_different_kind(self, mix, all_arithmetic_functions):
         op = all_arithmetic_functions
 
-        values = self._base([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
-        rvalues = self._base([2, np.nan, 2, 3, np.nan, 0, 1, 5, 2, np.nan])
+        values = np.array([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
+        rvalues = np.array([2, np.nan, 2, 3, np.nan, 0, 1, 5, 2, np.nan])
 
-        a = self._klass(values, kind="integer")
-        b = self._klass(rvalues, kind="block")
+        a = SparseArray(values, kind="integer")
+        b = SparseArray(rvalues, kind="block")
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
         self._check_numeric_ops(a, b * 0, values, rvalues * 0, mix, op)
 
-        a = self._klass(values, kind="integer", fill_value=0)
-        b = self._klass(rvalues, kind="block")
+        a = SparseArray(values, kind="integer", fill_value=0)
+        b = SparseArray(rvalues, kind="block")
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
-        a = self._klass(values, kind="integer", fill_value=0)
-        b = self._klass(rvalues, kind="block", fill_value=0)
+        a = SparseArray(values, kind="integer", fill_value=0)
+        b = SparseArray(rvalues, kind="block", fill_value=0)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
-        a = self._klass(values, kind="integer", fill_value=1)
-        b = self._klass(rvalues, kind="block", fill_value=2)
+        a = SparseArray(values, kind="integer", fill_value=1)
+        b = SparseArray(rvalues, kind="block", fill_value=2)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
     def test_float_array_comparison(self, kind):
-        values = self._base([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
-        rvalues = self._base([2, np.nan, 2, 3, np.nan, 0, 1, 5, 2, np.nan])
+        values = np.array([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
+        rvalues = np.array([2, np.nan, 2, 3, np.nan, 0, 1, 5, 2, np.nan])
 
-        a = self._klass(values, kind=kind)
-        b = self._klass(rvalues, kind=kind)
+        a = SparseArray(values, kind=kind)
+        b = SparseArray(rvalues, kind=kind)
         self._check_comparison_ops(a, b, values, rvalues)
         self._check_comparison_ops(a, b * 0, values, rvalues * 0)
 
-        a = self._klass(values, kind=kind, fill_value=0)
-        b = self._klass(rvalues, kind=kind)
+        a = SparseArray(values, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, kind=kind)
         self._check_comparison_ops(a, b, values, rvalues)
 
-        a = self._klass(values, kind=kind, fill_value=0)
-        b = self._klass(rvalues, kind=kind, fill_value=0)
+        a = SparseArray(values, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, kind=kind, fill_value=0)
         self._check_comparison_ops(a, b, values, rvalues)
 
-        a = self._klass(values, kind=kind, fill_value=1)
-        b = self._klass(rvalues, kind=kind, fill_value=2)
+        a = SparseArray(values, kind=kind, fill_value=1)
+        b = SparseArray(rvalues, kind=kind, fill_value=2)
         self._check_comparison_ops(a, b, values, rvalues)
 
     def test_int_array(self, kind, mix, all_arithmetic_functions):
@@ -276,33 +272,33 @@ class TestSparseArrayArithmetics:
         # have to specify dtype explicitly until fixing GH 667
         dtype = np.int64
 
-        values = self._base([0, 1, 2, 0, 0, 0, 1, 2, 1, 0], dtype=dtype)
-        rvalues = self._base([2, 0, 2, 3, 0, 0, 1, 5, 2, 0], dtype=dtype)
+        values = np.array([0, 1, 2, 0, 0, 0, 1, 2, 1, 0], dtype=dtype)
+        rvalues = np.array([2, 0, 2, 3, 0, 0, 1, 5, 2, 0], dtype=dtype)
 
-        a = self._klass(values, dtype=dtype, kind=kind)
+        a = SparseArray(values, dtype=dtype, kind=kind)
         assert a.dtype == SparseDtype(dtype)
-        b = self._klass(rvalues, dtype=dtype, kind=kind)
+        b = SparseArray(rvalues, dtype=dtype, kind=kind)
         assert b.dtype == SparseDtype(dtype)
 
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
         self._check_numeric_ops(a, b * 0, values, rvalues * 0, mix, op)
 
-        a = self._klass(values, fill_value=0, dtype=dtype, kind=kind)
+        a = SparseArray(values, fill_value=0, dtype=dtype, kind=kind)
         assert a.dtype == SparseDtype(dtype)
-        b = self._klass(rvalues, dtype=dtype, kind=kind)
+        b = SparseArray(rvalues, dtype=dtype, kind=kind)
         assert b.dtype == SparseDtype(dtype)
 
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
-        a = self._klass(values, fill_value=0, dtype=dtype, kind=kind)
+        a = SparseArray(values, fill_value=0, dtype=dtype, kind=kind)
         assert a.dtype == SparseDtype(dtype)
-        b = self._klass(rvalues, fill_value=0, dtype=dtype, kind=kind)
+        b = SparseArray(rvalues, fill_value=0, dtype=dtype, kind=kind)
         assert b.dtype == SparseDtype(dtype)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
-        a = self._klass(values, fill_value=1, dtype=dtype, kind=kind)
+        a = SparseArray(values, fill_value=1, dtype=dtype, kind=kind)
         assert a.dtype == SparseDtype(dtype, fill_value=1)
-        b = self._klass(rvalues, fill_value=2, dtype=dtype, kind=kind)
+        b = SparseArray(rvalues, fill_value=2, dtype=dtype, kind=kind)
         assert b.dtype == SparseDtype(dtype, fill_value=2)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
@@ -310,46 +306,46 @@ class TestSparseArrayArithmetics:
         dtype = "int64"
         # int32 NI ATM
 
-        values = self._base([0, 1, 2, 0, 0, 0, 1, 2, 1, 0], dtype=dtype)
-        rvalues = self._base([2, 0, 2, 3, 0, 0, 1, 5, 2, 0], dtype=dtype)
+        values = np.array([0, 1, 2, 0, 0, 0, 1, 2, 1, 0], dtype=dtype)
+        rvalues = np.array([2, 0, 2, 3, 0, 0, 1, 5, 2, 0], dtype=dtype)
 
-        a = self._klass(values, dtype=dtype, kind=kind)
-        b = self._klass(rvalues, dtype=dtype, kind=kind)
+        a = SparseArray(values, dtype=dtype, kind=kind)
+        b = SparseArray(rvalues, dtype=dtype, kind=kind)
         self._check_comparison_ops(a, b, values, rvalues)
         self._check_comparison_ops(a, b * 0, values, rvalues * 0)
 
-        a = self._klass(values, dtype=dtype, kind=kind, fill_value=0)
-        b = self._klass(rvalues, dtype=dtype, kind=kind)
+        a = SparseArray(values, dtype=dtype, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, dtype=dtype, kind=kind)
         self._check_comparison_ops(a, b, values, rvalues)
 
-        a = self._klass(values, dtype=dtype, kind=kind, fill_value=0)
-        b = self._klass(rvalues, dtype=dtype, kind=kind, fill_value=0)
+        a = SparseArray(values, dtype=dtype, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, dtype=dtype, kind=kind, fill_value=0)
         self._check_comparison_ops(a, b, values, rvalues)
 
-        a = self._klass(values, dtype=dtype, kind=kind, fill_value=1)
-        b = self._klass(rvalues, dtype=dtype, kind=kind, fill_value=2)
+        a = SparseArray(values, dtype=dtype, kind=kind, fill_value=1)
+        b = SparseArray(rvalues, dtype=dtype, kind=kind, fill_value=2)
         self._check_comparison_ops(a, b, values, rvalues)
 
     @pytest.mark.parametrize("fill_value", [True, False, np.nan])
     def test_bool_same_index(self, kind, fill_value):
         # GH 14000
         # when sp_index are the same
-        values = self._base([True, False, True, True], dtype=np.bool_)
-        rvalues = self._base([True, False, True, True], dtype=np.bool_)
+        values = np.array([True, False, True, True], dtype=np.bool_)
+        rvalues = np.array([True, False, True, True], dtype=np.bool_)
 
-        a = self._klass(values, kind=kind, dtype=np.bool_, fill_value=fill_value)
-        b = self._klass(rvalues, kind=kind, dtype=np.bool_, fill_value=fill_value)
+        a = SparseArray(values, kind=kind, dtype=np.bool_, fill_value=fill_value)
+        b = SparseArray(rvalues, kind=kind, dtype=np.bool_, fill_value=fill_value)
         self._check_logical_ops(a, b, values, rvalues)
 
     @pytest.mark.parametrize("fill_value", [True, False, np.nan])
     def test_bool_array_logical(self, kind, fill_value):
         # GH 14000
         # when sp_index are the same
-        values = self._base([True, False, True, False, True, True], dtype=np.bool_)
-        rvalues = self._base([True, False, False, True, False, True], dtype=np.bool_)
+        values = np.array([True, False, True, False, True, True], dtype=np.bool_)
+        rvalues = np.array([True, False, False, True, False, True], dtype=np.bool_)
 
-        a = self._klass(values, kind=kind, dtype=np.bool_, fill_value=fill_value)
-        b = self._klass(rvalues, kind=kind, dtype=np.bool_, fill_value=fill_value)
+        a = SparseArray(values, kind=kind, dtype=np.bool_, fill_value=fill_value)
+        b = SparseArray(rvalues, kind=kind, dtype=np.bool_, fill_value=fill_value)
         self._check_logical_ops(a, b, values, rvalues)
 
     def test_mixed_array_float_int(self, kind, mix, all_arithmetic_functions, request):
@@ -361,28 +357,28 @@ class TestSparseArrayArithmetics:
 
         rdtype = "int64"
 
-        values = self._base([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
-        rvalues = self._base([2, 0, 2, 3, 0, 0, 1, 5, 2, 0], dtype=rdtype)
+        values = np.array([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
+        rvalues = np.array([2, 0, 2, 3, 0, 0, 1, 5, 2, 0], dtype=rdtype)
 
-        a = self._klass(values, kind=kind)
-        b = self._klass(rvalues, kind=kind)
+        a = SparseArray(values, kind=kind)
+        b = SparseArray(rvalues, kind=kind)
         assert b.dtype == SparseDtype(rdtype)
 
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
         self._check_numeric_ops(a, b * 0, values, rvalues * 0, mix, op)
 
-        a = self._klass(values, kind=kind, fill_value=0)
-        b = self._klass(rvalues, kind=kind)
+        a = SparseArray(values, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, kind=kind)
         assert b.dtype == SparseDtype(rdtype)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
-        a = self._klass(values, kind=kind, fill_value=0)
-        b = self._klass(rvalues, kind=kind, fill_value=0)
+        a = SparseArray(values, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, kind=kind, fill_value=0)
         assert b.dtype == SparseDtype(rdtype)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
-        a = self._klass(values, kind=kind, fill_value=1)
-        b = self._klass(rvalues, kind=kind, fill_value=2)
+        a = SparseArray(values, kind=kind, fill_value=1)
+        b = SparseArray(rvalues, kind=kind, fill_value=2)
         assert b.dtype == SparseDtype(rdtype, fill_value=2)
         self._check_numeric_ops(a, b, values, rvalues, mix, op)
 
@@ -390,28 +386,28 @@ class TestSparseArrayArithmetics:
         rdtype = "int64"
         # int32 NI ATM
 
-        values = self._base([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
-        rvalues = self._base([2, 0, 2, 3, 0, 0, 1, 5, 2, 0], dtype=rdtype)
+        values = np.array([np.nan, 1, 2, 0, np.nan, 0, 1, 2, 1, np.nan])
+        rvalues = np.array([2, 0, 2, 3, 0, 0, 1, 5, 2, 0], dtype=rdtype)
 
-        a = self._klass(values, kind=kind)
-        b = self._klass(rvalues, kind=kind)
+        a = SparseArray(values, kind=kind)
+        b = SparseArray(rvalues, kind=kind)
         assert b.dtype == SparseDtype(rdtype)
 
         self._check_comparison_ops(a, b, values, rvalues)
         self._check_comparison_ops(a, b * 0, values, rvalues * 0)
 
-        a = self._klass(values, kind=kind, fill_value=0)
-        b = self._klass(rvalues, kind=kind)
+        a = SparseArray(values, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, kind=kind)
         assert b.dtype == SparseDtype(rdtype)
         self._check_comparison_ops(a, b, values, rvalues)
 
-        a = self._klass(values, kind=kind, fill_value=0)
-        b = self._klass(rvalues, kind=kind, fill_value=0)
+        a = SparseArray(values, kind=kind, fill_value=0)
+        b = SparseArray(rvalues, kind=kind, fill_value=0)
         assert b.dtype == SparseDtype(rdtype)
         self._check_comparison_ops(a, b, values, rvalues)
 
-        a = self._klass(values, kind=kind, fill_value=1)
-        b = self._klass(rvalues, kind=kind, fill_value=2)
+        a = SparseArray(values, kind=kind, fill_value=1)
+        b = SparseArray(rvalues, kind=kind, fill_value=2)
         assert b.dtype == SparseDtype(rdtype, fill_value=2)
         self._check_comparison_ops(a, b, values, rvalues)
 
@@ -495,36 +491,60 @@ def test_sparray_inplace():
     tm.assert_sp_array_equal(sparray, expected)
 
 
-@pytest.mark.parametrize("fill_value", [True, False])
-def test_invert(fill_value):
-    arr = np.array([True, False, False, True])
-    sparray = SparseArray(arr, fill_value=fill_value)
-    result = ~sparray
-    expected = SparseArray(~arr, fill_value=not fill_value)
-    tm.assert_sp_array_equal(result, expected)
-
-    result = ~pd.Series(sparray)
-    expected = pd.Series(expected)
-    tm.assert_series_equal(result, expected)
-
-    result = ~pd.DataFrame({"A": sparray})
-    expected = pd.DataFrame({"A": expected})
-    tm.assert_frame_equal(result, expected)
-
-
-@pytest.mark.parametrize("fill_value", [0, np.nan])
-@pytest.mark.parametrize("op", [operator.pos, operator.neg])
-def test_unary_op(op, fill_value):
-    arr = np.array([0, 1, np.nan, 2])
-    sparray = SparseArray(arr, fill_value=fill_value)
-    result = op(sparray)
-    expected = SparseArray(op(arr), fill_value=op(fill_value))
-    tm.assert_sp_array_equal(result, expected)
-
-
 @pytest.mark.parametrize("cons", [list, np.array, SparseArray])
 def test_mismatched_length_cmp_op(cons):
     left = SparseArray([True, True])
     right = cons([True, True, True])
     with pytest.raises(ValueError, match="operands have mismatched length"):
         left & right
+
+
+@pytest.mark.parametrize("op", ["add", "sub", "mul", "truediv", "floordiv", "pow"])
+def test_binary_operators(op):
+    op = getattr(operator, op)
+    data1 = np.random.randn(20)
+    data2 = np.random.randn(20)
+
+    data1[::2] = np.nan
+    data2[::3] = np.nan
+
+    arr1 = SparseArray(data1)
+    arr2 = SparseArray(data2)
+
+    data1[::2] = 3
+    data2[::3] = 3
+    farr1 = SparseArray(data1, fill_value=3)
+    farr2 = SparseArray(data2, fill_value=3)
+
+    def _check_op(op, first, second):
+        res = op(first, second)
+        exp = SparseArray(
+            op(first.to_dense(), second.to_dense()), fill_value=first.fill_value
+        )
+        assert isinstance(res, SparseArray)
+        tm.assert_almost_equal(res.to_dense(), exp.to_dense())
+
+        res2 = op(first, second.to_dense())
+        assert isinstance(res2, SparseArray)
+        tm.assert_sp_array_equal(res, res2)
+
+        res3 = op(first.to_dense(), second)
+        assert isinstance(res3, SparseArray)
+        tm.assert_sp_array_equal(res, res3)
+
+        res4 = op(first, 4)
+        assert isinstance(res4, SparseArray)
+
+        # Ignore this if the actual op raises (e.g. pow).
+        try:
+            exp = op(first.to_dense(), 4)
+            exp_fv = op(first.fill_value, 4)
+        except ValueError:
+            pass
+        else:
+            tm.assert_almost_equal(res4.fill_value, exp_fv)
+            tm.assert_almost_equal(res4.to_dense(), exp)
+
+    with np.errstate(all="ignore"):
+        for first_arr, second_arr in [(arr1, arr2), (farr1, farr2)]:
+            _check_op(op, first_arr, second_arr)
